fix(lsp): completion continues working on subsequent requests
Problem: When isIncomplete is true, subsequent input will continue to send requests. The prefix is always taken from the column where the request was triggered, which causes matches to be empty. and there is missing handle item.textEdit.

Solution: if word not match using filterText as word and accept textEdit, When Context.cusror have value use it for prefix sub