fix(lsp): enable insertReplaceSupport for use in adjust_start_col
Currently, with the typescript lsps typescript-language-server and
vtsls, omnicompletion on partial tokens for certain types, such as array
methods, and functions that are attached as attributes to other
functions, either results in no entries populated in the completion menu
(typescript-language-server), or an unfiltered completion menu with all
array methods included, even if they don't share the same prefix as the
partial token being completed (vtsls).

This change enables insertReplaceSupport and uses the insert portion of
the lsp completion response in adjust_start_col if it's included in the
response. This fixes the issues above.

Completion results are still filtered client side.